requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
httpx>=0.24.0
pydantic>=2.0.0
langdetect>=1.0.9
python-dotenv>=1.0.0
requests>=2.28.0
aiofiles>=23.2.1
motor>=3.1.1
pydantic-ai[openai]>=0.2.0
mem0ai>=1.0.0
chromadb>=0.4.0
openai>=1.0.0